refactor: rename package routes
