+++ internal/api/routes/error_handler.go
@@ -1,4 +1,4 @@
-package routeshandlers
+package routes
 
 import (
 	"github.com/gin-gonic/gin"
