+++ internal/api/routes/ping_route.go
@@ -1,4 +1,4 @@
-package routeshandlers
+package routes
 
 import (
 	"github.com/gin-gonic/gin"
