+++ internal/api/routes/setup_router.go
@@ -1,4 +1,4 @@
-package routeshandlers
+package routes
 
 import (
 	"github.com/GuillaumeDeconinck/todos-go/internal/api/configuration"
